player changes
- move player globals out of players header
- move some other structs into their correct locations
